requirement.txt

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML & NLP
torch==2.1.0
transformers==4.35.0
spacy==3.7.2
datasets==2.14.6
scikit-learn==1.3.2
seqeval==1.2.2

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.1.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
python-dateutil==2.8.2

# MLOps & Monitoring
mlflow==2.8.1
prometheus-client==0.19.0
python-json-logger==2.0.7
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1
faker==20.1.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
EOF